backend/requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.0
websockets==12.0
python-multipart==0.0.9
aiosqlite==0.20.0
pydantic==2.9.0
python-dotenv==1.0.1
aiohttp==3.10.0

# Report Generation
reportlab==4.2.0
openpyxl==3.1.2
python-docx==1.1.0